Changed int input to accept 64 bits

+++ src/engine/source/builder/builders/opBuilderSCAdecoder.hpp
@@ -117,7 +117,7 @@ struct DecodeCxt
      * @param field Field to get the value.
      * @return empty if the field is not found or the value is not an int
      */
-    std::optional<int> getSrcInt(sca::field::Name field) const { return event->getInt(sourcePath.at(field)); };
+    std::optional<int> getSrcInt(sca::field::Name field) const { return event->getInt64(sourcePath.at(field)); };
 
     /**
      * @brief Get number as a double value from a field.
